BlockStoreStorageAPI is not separate lib component
