+++ all_sources.bat
@@ -3,9 +3,11 @@ set LIB_THIRDPARTY_SRC=%BASE_DIR%src\ext\*.c
 
 set ATOMICCANCEL_SRC=%BASE_DIR%lib\atomiccancel\*.c
 
+set COMPRESSBLOCKSTORE_SRC=%BASE_DIR%lib\compressblockstore\*.c
+
 set CACHEBLOCKSTORE_SRC=%BASE_DIR%lib\cacheblockstore\*.c
 
-set COMPRESSBLOCKSTORE_SRC=%BASE_DIR%lib\compressblockstore\*.c
+set BLOCKSTORESTORAGE_SRC=%BASE_DIR%lib\blockstorestorage\*.c
 
 set FILESTORAGE_SRC=%BASE_DIR%lib\*.c
 
@@ -42,7 +44,7 @@ set BROTLI_THIRDPARTY_SRC=%BASE_DIR%lib\brotli\ext\common\*.c %BASE_DIR%lib\brot
 set ZSTD_SRC=%BASE_DIR%lib\zstd\*.c
 set ZSTD_THIRDPARTY_SRC=%BASE_DIR%lib\zstd\ext\common\*.c %BASE_DIR%lib\zstd\ext\compress\*.c %BASE_DIR%lib\zstd\ext\decompress\*.c
 
-set SRC=%BASE_DIR%src\*.c %LIB_SRC% %ATOMICCANCEL_SRC% %CACHEBLOCKSTORE_SRC% %COMPRESSBLOCKSTORE_SRC% %RETAININGBLOCKSTORE_SRC% %SHAREBLOCKSTORE_SRC% %FILESTORAGE_SRC% %FSBLOCKSTORE_SRC% %MEMSTORAGE_SRC% %MEOWHASH_SRC% %COMPRESSION_REGISTRY_SRC% %HASH_REGISTRY_SRC% %BIKESHED_SRC% %BLAKE2_SRC% %BLAKE3_SRC% %LZ4_SRC% %BROTLI_SRC% %ZSTD_SRC%
+set SRC=%BASE_DIR%src\*.c %LIB_SRC% %ATOMICCANCEL_SRC% %COMPRESSBLOCKSTORE_SRC% %CACHEBLOCKSTORE_SRC% %BLOCKSTORESTORAGE_SRC% %RETAININGBLOCKSTORE_SRC% %SHAREBLOCKSTORE_SRC% %FILESTORAGE_SRC% %FSBLOCKSTORE_SRC% %MEMSTORAGE_SRC% %MEOWHASH_SRC% %COMPRESSION_REGISTRY_SRC% %HASH_REGISTRY_SRC% %BIKESHED_SRC% %BLAKE2_SRC% %BLAKE3_SRC% %LZ4_SRC% %BROTLI_SRC% %ZSTD_SRC%
 set THIRDPARTY_SRC=%LIB_THIRDPARTY_SRC% %BLAKE2_THIRDPARTY_SRC% %BLAKE3_THIRDPARTY_SRC% %LZ4_THIRDPARTY_SRC% %BROTLI_THIRDPARTY_SRC% %ZSTD_THIRDPARTY_SRC%
 set THIRDPARTY_SRC_SSE42=
 set THIRDPARTY_SRC_AVX2=%BLAKE3_THIRDPARTY_AVX2%
