+++ all_sources.sh
@@ -5,6 +5,8 @@ LIB_THIRDPARTY_SRC="${BASE_DIR}src/ext/*.c"
 
 ATOMICCANCEL_SRC="${BASE_DIR}lib/atomiccancel/*.c"
 
+COMPRESSBLOCKSTORE_SRC="${BASE_DIR}lib/compressblockstore/*.c"
+
 CACHEBLOCKSTORE_SRC="${BASE_DIR}lib/cacheblockstore/*.c"
 
 COMPRESSBLOCKSTORE_SRC="${BASE_DIR}lib/compressblockstore/*.c"
@@ -44,7 +46,7 @@ BROTLI_THIRDPARTY_SRC="${BASE_DIR}lib/brotli/ext/common/*.c ${BASE_DIR}lib/brotl
 ZSTD_SRC="${BASE_DIR}lib/zstd/*.c"
 ZSTD_THIRDPARTY_SRC="${BASE_DIR}lib/zstd/ext/common/*.c ${BASE_DIR}lib/zstd/ext/compress/*.c ${BASE_DIR}lib/zstd/ext/decompress/*.c"
 
-export SRC="${BASE_DIR}src/*.c $LIB_SRC $ATOMICCANCEL_SRC $CACHEBLOCKSTORE_SRC $COMPRESSBLOCKSTORE_SRC $RETAININGBLOCKSTORE_SRC $SHAREBLOCKSTORE_SRC $FILESTORAGE_SRC $FSBLOCKSTORAGE_SRC $MEMSTORAGE_SRC $MEOWHASH_SRC $COMPRESSION_REGISTRY_SRC $HASH_REGISTRY_SRC $BIKESHED_SRC $BLAKE2_SRC $BLAKE3_SRC $LZ4_SRC $BROTLI_SRC $ZSTD_SRC"
+export SRC="${BASE_DIR}src/*.c $LIB_SRC $ATOMICCANCEL_SRC $COMPRESSBLOCKSTORE_SRC $CACHEBLOCKSTORE_SRC $COMPRESSBLOCKSTORE_SRC $RETAININGBLOCKSTORE_SRC $SHAREBLOCKSTORE_SRC $FILESTORAGE_SRC $FSBLOCKSTORAGE_SRC $MEMSTORAGE_SRC $MEOWHASH_SRC $COMPRESSION_REGISTRY_SRC $HASH_REGISTRY_SRC $BIKESHED_SRC $BLAKE2_SRC $BLAKE3_SRC $LZ4_SRC $BROTLI_SRC $ZSTD_SRC"
 export THIRDPARTY_SRC="$LIB_THIRDPARTY_SRC $BLAKE2_THIRDPARTY_SRC $BLAKE3_THIRDPARTY_SRC $LZ4_THIRDPARTY_SRC $BROTLI_THIRDPARTY_SRC $ZSTD_THIRDPARTY_SRC"
 export THIRDPARTY_SRC_SSE42=""
 export THIRDPARTY_SRC_AVX2="$BLAKE3_THIRDPARTY_AVX2"
