+++ test/test.cpp
@@ -8,6 +8,7 @@
 #include "../lib/bikeshed/longtail_bikeshed.h"
 #include "../lib/brotli/longtail_brotli.h"
 #include "../lib/atomiccancel/longtail_atomiccancel.h"
+#include "../lib/blockstorestorage/longtail_blockstorestorage.h"
 #include "../lib/cacheblockstore/longtail_cacheblockstore.h"
 #include "../lib/compressblockstore/longtail_compressblockstore.h"
 #include "../lib/compressionregistry/longtail_full_compression_registry.h"
@@ -6108,779 +6109,6 @@ TEST(Longtail, TestChangeVersionDiskFull)
     SAFE_DISPOSE_API(mem_storage);
 }
 
-struct HashToIndexLookup
-{
-    TLongtail_Hash key;
-    uint64_t value;
-};
-
-struct BlockStoreFS
-{
-    struct Longtail_StorageAPI m_API;
-    Longtail_HashAPI* hash_api;
-    struct Longtail_JobAPI* job_api;
-    struct Longtail_BlockStoreAPI* block_store;
-    struct Longtail_ContentIndex* content_index;
-    struct Longtail_VersionIndex* version_index;
-    struct HashToIndexLookup* chunk_hash_to_block_index_lookup;
-    struct HashToIndexLookup* asset_lookup;
-    uint64_t* chunk_asset_offsets;
-};
-
-struct OpenBlockStoreFile
-{
-    uint32_t asset_index;
-    uint32_t seek_chunk_offset;
-    uint64_t seek_asset_pos;
-};
-
-int OpenAsset(struct BlockStoreFS* block_store_fs, const char* path, struct OpenBlockStoreFile** out_block_store_file)
-{
-    uint64_t path_hash = 0;
-    int err = Longtail_GetPathHash(block_store_fs->hash_api, path, &path_hash);
-    if (err)
-    {
-        return err;
-    }
-    intptr_t asset_index_ptr = hmgeti(block_store_fs->asset_lookup, path_hash);
-    if (asset_index_ptr == -1)
-    {
-        return ENOENT;
-    }
-    uint64_t asset_index = block_store_fs->asset_lookup[asset_index_ptr].value;
-    struct OpenBlockStoreFile* block_store_file = (struct OpenBlockStoreFile*)Longtail_Alloc(sizeof(struct OpenBlockStoreFile));
-    block_store_file->asset_index = (uint32_t)asset_index;
-    block_store_file->seek_chunk_offset = 0;
-    block_store_file->seek_asset_pos = 0;
-    *out_block_store_file = block_store_file;
-    return 0;
-}
-
-void CloseAsset(struct BlockStoreFS* block_store_fs, struct OpenBlockStoreFile* block_store_file)
-{
-    Longtail_Free(block_store_file);
-}
-
-int GetAssetSize(struct BlockStoreFS* block_store_fs, struct OpenBlockStoreFile* block_store_file, uint64_t* out_size)
-{
-    *out_size = block_store_fs->version_index->m_AssetSizes[block_store_file->asset_index];
-    return 0;
-}
-
-struct ChunkRange
-{
-    uint64_t asset_start_offset;
-    uint32_t chunk_start;
-    uint32_t chunk_end;
-};
-
-struct BlockRange
-{
-    TLongtail_Hash key;
-    struct ChunkRange value;
-};
-
-int ReadFromBlock(
-    struct Longtail_StoredBlock* stored_block,
-    struct ChunkRange range,
-    struct BlockStoreFS* block_store_fs,
-    struct OpenBlockStoreFile* block_store_file,
-    uint64_t start,
-    uint64_t size,
-    char* buffer,
-    const uint32_t* chunk_indexes)
-{
-    uint64_t read_end = start + size;
-    uint32_t chunk_block_offset = 0;
-    struct HashToIndexLookup* block_chunk_lookup = 0;
-    for (uint32_t c = 0; c < *stored_block->m_BlockIndex->m_ChunkCount; ++c)
-    {
-        TLongtail_Hash chunk_hash = stored_block->m_BlockIndex->m_ChunkHashes[c];
-        hmput(block_chunk_lookup, chunk_hash, chunk_block_offset);
-        chunk_block_offset += stored_block->m_BlockIndex->m_ChunkSizes[c];
-    }
-    uint64_t asset_offset = range.asset_start_offset;
-    for (uint32_t c = range.chunk_start; c < range.chunk_end; ++c)
-    {
-        uint32_t chunk_index = chunk_indexes[c];
-        TLongtail_Hash chunk_hash = block_store_fs->version_index->m_ChunkHashes[chunk_index];
-        uint32_t chunk_size = block_store_fs->version_index->m_ChunkSizes[chunk_index];
-        uint64_t asset_offset_chunk_end = asset_offset + chunk_size;
-        if (asset_offset_chunk_end < start)
-        {
-            // We sure could use a more efficient lookup to pos here :)
-            asset_offset = asset_offset_chunk_end;
-            continue;
-        }
-
-        const char* block_data = (char*)stored_block->m_BlockData;
-
-        intptr_t chunk_block_offset_ptr = hmgeti(block_chunk_lookup, chunk_hash);
-        if (chunk_block_offset_ptr == -1)
-        {
-            asset_offset += chunk_size;
-            continue;
-        }
-        uint32_t chunk_offset = 0;
-        uint32_t read_length = chunk_size;
-        if (asset_offset < start)
-        {
-            chunk_offset = (uint32_t)(start - asset_offset);
-            read_length = chunk_size - chunk_offset;
-            asset_offset = start;
-        }
-        if (asset_offset + read_length > read_end)
-        {
-            read_length = (uint32_t)(read_end - asset_offset);
-        }
-
-        uint32_t chunk_block_offset = (uint32_t)block_chunk_lookup[chunk_block_offset_ptr].value;
-        memcpy(&buffer[asset_offset - start], &block_data[chunk_block_offset + chunk_offset], read_length);
-        asset_offset += read_length;
-    }
-    hmfree(block_chunk_lookup);
-    return 0;
-}
-
-struct ReadFromBlockJobData
-{
-    uint64_t block_index;
-    struct ChunkRange range;
-    struct BlockStoreFS* block_store_fs;
-    struct OpenBlockStoreFile* block_store_file;
-    uint64_t start;
-    uint64_t size;
-    char* buffer;
-    const uint32_t* chunk_indexes;
-    struct Longtail_StoredBlock* stored_block;
-    int err;
-};
-
-struct TestReadBlockComplete
-{
-    struct Longtail_AsyncGetStoredBlockAPI m_API;
-    uint32_t job_id;
-    struct ReadFromBlockJobData* data;
-    static void OnComplete(struct Longtail_AsyncGetStoredBlockAPI* async_complete_api, Longtail_StoredBlock* stored_block, int err)
-    {
-        struct TestReadBlockComplete* cb = (struct TestReadBlockComplete*)async_complete_api;
-        struct Longtail_JobAPI* job_api = cb->data->block_store_fs->job_api;
-        cb->data->err = err;
-        cb->data->stored_block = stored_block;
-        job_api->ResumeJob(job_api, cb->job_id);
-        Longtail_Free(cb);
-    }
-};
-
-
-int ReadFromBlockJob(void* context, uint32_t job_id, int is_cancelled)
-{
-    struct ReadFromBlockJobData* data = (struct ReadFromBlockJobData*)context;
-
-    if (!data->stored_block)
-    {
-        if (data->err)
-        {
-            return 0;
-        }
-        // Don't need dynamic alloc, could be part of struct ReadFromBlockJobData since it covers the lifetime of struct TestReadBlockComplete
-        struct TestReadBlockComplete* complete_cb = (struct TestReadBlockComplete*)Longtail_Alloc(sizeof(struct TestReadBlockComplete));
-        complete_cb->m_API.m_API.Dispose = 0;
-        complete_cb->m_API.OnComplete = TestReadBlockComplete::OnComplete;
-        complete_cb->job_id = job_id;
-        complete_cb->data = data;
-
-        TLongtail_Hash block_hash = data->block_store_fs->content_index->m_BlockHashes[data->block_index];
-        int err = data->block_store_fs->block_store->GetStoredBlock(data->block_store_fs->block_store, block_hash, &complete_cb->m_API);
-        if (err)
-        {
-            data->err = err;
-            Longtail_Free(complete_cb);
-            return 0;
-        }
-        return EBUSY;
-    }
-
-    data->err = ReadFromBlock(
-        data->stored_block,
-        data->range,
-        data->block_store_fs,
-        data->block_store_file,
-        data->start,
-        data->size,
-        data->buffer,
-        data->chunk_indexes);
-    if (data->stored_block->Dispose)
-    {
-        data->stored_block->Dispose(data->stored_block);
-    }
-    return 0;
-}
-
-static uint32_t FindStartChunk(const uint64_t* a, uint32_t n, uint64_t val) {
-    uint32_t first = 0;
-    uint32_t count = n;
-    while (count > 0)
-    {
-        uint32_t step = count/2;
-        uint32_t it = first + step;
-        if (a[it] < val)
-        {
-            first = it + 1;
-            count -= step + 1;
-        }
-        else
-        {
-            count = step;
-        }
-    }
-    if (first == n)
-    {
-        return first - 1;
-    }
-    if (a[first] > val)
-    {
-        return first - 1;
-    }
-    return first;
-}
-
-int Seek(struct BlockStoreFS* block_store_fs,
-    struct OpenBlockStoreFile* block_store_file,
-    uint64_t pos)
-{
-    if (pos == block_store_file->seek_asset_pos)
-    {
-        return 0;
-    }
-    uint32_t chunk_count = block_store_fs->version_index->m_AssetChunkCounts[block_store_file->asset_index];
-    if (chunk_count == 0)
-    {
-        return pos == 0 ? 0 : EIO;
-    }
-    uint32_t chunk_start_index = block_store_fs->version_index->m_AssetChunkIndexStarts[block_store_file->asset_index];
-
-    const uint64_t* chunk_asset_offsets = &block_store_fs->chunk_asset_offsets[chunk_start_index];
-    if (block_store_file->seek_chunk_offset < chunk_count)
-    {
-        if (pos >= chunk_asset_offsets[block_store_file->seek_chunk_offset])
-        {
-            if (block_store_file->seek_chunk_offset == chunk_count - 1)
-            {
-                // on last block
-                return 0;
-            }
-            else if (pos < chunk_asset_offsets[block_store_file->seek_chunk_offset + 1])
-            {
-                return 0;
-            }
-        }
-    }
-
-    const uint32_t* chunk_indexes = &block_store_fs->version_index->m_AssetChunkIndexes[chunk_start_index];
-
-    uint64_t asset_size = block_store_fs->version_index->m_AssetSizes[block_store_file->asset_index];
-
-    uint32_t start_chunk_index = FindStartChunk(chunk_asset_offsets, chunk_count, pos);
-    uint64_t start_asset_offset = chunk_asset_offsets[start_chunk_index];
-
-    block_store_file->seek_asset_pos = start_asset_offset;
-    block_store_file->seek_chunk_offset = start_chunk_index;
-    return 0;
-}
-
-int ReadBlockStoreFile(
-    struct BlockStoreFS* block_store_fs,
-    struct OpenBlockStoreFile* block_store_file,
-    uint64_t start,
-    uint64_t size,
-    void* out_buffer)
-{
-    char* buffer = (char*)out_buffer;
-
-    uint32_t chunk_count = block_store_fs->version_index->m_AssetChunkCounts[block_store_file->asset_index];
-    uint32_t chunk_start_index = block_store_fs->version_index->m_AssetChunkIndexStarts[block_store_file->asset_index];
-    const uint32_t* chunk_indexes = &block_store_fs->version_index->m_AssetChunkIndexes[chunk_start_index];
-
-    const uint64_t read_end = start + size;
-    int err = Seek(block_store_fs, block_store_file, start);
-    if (err)
-    {
-        return err;
-    }
-
-    uint32_t seek_chunk_offset = block_store_file->seek_chunk_offset;
-    uint64_t seek_asset_pos = block_store_file->seek_asset_pos;
-
-    struct BlockRange* block_range_map = 0;
-    uint64_t* block_indexes = 0;
-
-    intptr_t last_block_index = (uint64_t)-1;
-    uint64_t last_block_range_index_ptr = -1;
-
-    for (uint32_t c = seek_chunk_offset; c < chunk_count; ++c)
-    {
-        uint32_t chunk_index = chunk_indexes[c];
-        TLongtail_Hash chunk_hash = block_store_fs->version_index->m_ChunkHashes[chunk_index];
-        intptr_t block_find_index_ptr = hmgeti(block_store_fs->chunk_hash_to_block_index_lookup, chunk_hash);
-        uint64_t block_index = block_store_fs->chunk_hash_to_block_index_lookup[block_find_index_ptr].value;
-        if ((block_index != (uint64_t)-1) && (block_index == last_block_index))
-        {
-            block_range_map[last_block_range_index_ptr].value.chunk_end = c + 1;
-        }
-        else
-        {
-            last_block_range_index_ptr = hmgeti(block_range_map, block_index);
-            if (last_block_range_index_ptr == -1)
-            {
-                struct ChunkRange range = {seek_asset_pos, c, c + 1};
-                hmput(block_range_map, block_index, range);
-                arrput(block_indexes, block_index);
-                last_block_range_index_ptr = hmgeti(block_range_map, block_index);
-            }
-            else
-            {
-                block_range_map[last_block_range_index_ptr].value.chunk_end = c + 1;
-            }
-            last_block_index = block_index;
-        }
-        block_store_file->seek_chunk_offset = c;
-        block_store_file->seek_asset_pos = seek_asset_pos;
-
-        uint32_t chunk_size = block_store_fs->version_index->m_ChunkSizes[chunk_index];
-        seek_asset_pos += chunk_size;
-        if (seek_asset_pos >= read_end)
-        {
-            break;
-        }
-    }
-
-    intptr_t block_count = arrlen(block_indexes);
-    if (block_count == 0)
-    {
-        return 0;
-    }
-
-    if (block_count == 1)
-    {
-        uint64_t block_index = block_indexes[0];
-        struct ChunkRange range = hmget(block_range_map, block_index);
-
-        TLongtail_Hash block_hash = block_store_fs->content_index->m_BlockHashes[block_index];
-        TestAsyncGetBlockComplete getCB;
-        int err = block_store_fs->block_store->GetStoredBlock(block_store_fs->block_store, block_hash, &getCB.m_API);
-        if (err)
-        {
-            return err;
-        }
-        if (getCB.m_Err)
-        {
-            return getCB.m_Err;
-        }
-        struct Longtail_StoredBlock* stored_block = getCB.m_StoredBlock;
-
-        err = ReadFromBlock(
-            stored_block,
-            range,
-            block_store_fs,
-            block_store_file,
-            start,
-            size,
-            buffer,
-            chunk_indexes);
-        if (stored_block->Dispose)
-        {
-            stored_block->Dispose(stored_block);
-        }
-        if (err)
-        {
-            return err;
-        }
-    }
-    else
-    {
-        struct Longtail_JobAPI* job_api = block_store_fs->job_api;
-        uint32_t block_count = (uint32_t)arrlen(block_indexes);
-        struct ReadFromBlockJobData* job_datas = (struct ReadFromBlockJobData*)Longtail_Alloc(sizeof(struct ReadFromBlockJobData) * block_count);
-        LONGTAIL_FATAL_ASSERT(job_datas, return ENOMEM)
-        Longtail_JobAPI_Group job_group;
-        int err = job_api->ReserveJobs(job_api, block_count, &job_group);
-        LONGTAIL_FATAL_ASSERT(err == 0, return err)
-
-        Longtail_JobAPI_JobFunc* funcs = (Longtail_JobAPI_JobFunc*)Longtail_Alloc(sizeof(Longtail_JobAPI_JobFunc) * block_count);
-        void** ctxs = (void**)Longtail_Alloc(sizeof(void*) * block_count);
-
-        for (uint32_t b = 0; b < block_count; ++b)
-        {
-            uint64_t block_index = block_indexes[b];
-            struct ChunkRange range = hmget(block_range_map, block_index);
-
-            job_datas[b].block_index = block_index;
-            job_datas[b].range = range;
-            job_datas[b].block_store_fs = block_store_fs;
-            job_datas[b].block_store_file = block_store_file;
-            job_datas[b].start = start;
-            job_datas[b].size = size;
-            job_datas[b].buffer = buffer;
-            job_datas[b].chunk_indexes = chunk_indexes;
-            job_datas[b].stored_block = 0;
-            job_datas[b].err = 0;
-
-            funcs[b] = ReadFromBlockJob;
-            ctxs[b] = &job_datas[b];
-        }
-        Longtail_JobAPI_Jobs jobs;
-        err = job_api->CreateJobs(job_api, job_group, block_count, funcs, ctxs, &jobs);
-        LONGTAIL_FATAL_ASSERT(err == 0, return err)
-        err = job_api->ReadyJobs(job_api, block_count, jobs);
-        LONGTAIL_FATAL_ASSERT(err == 0, return err)
-        err = job_api->WaitForAllJobs(job_api, job_group, 0, 0, 0);
-        LONGTAIL_FATAL_ASSERT(err == 0, return err)
-        Longtail_Free(ctxs);
-        Longtail_Free(funcs);
-        Longtail_Free(job_datas);
-    }
-
-    hmfree(block_range_map);
-    arrfree(block_indexes);
-    return 0;
-}
-
-int Longtail_BlockStorageAPI_OpenReadFile(struct Longtail_StorageAPI* storage_api, const char* path, Longtail_StorageAPI_HOpenFile* out_open_file)
-{
-    struct BlockStoreFS* block_store_fs = (struct BlockStoreFS*)storage_api;
-    struct OpenBlockStoreFile* block_store_file;
-    int err = OpenAsset(block_store_fs, path, &block_store_file);
-    if (err)
-    {
-        return err;
-    }
-    *out_open_file = (Longtail_StorageAPI_HOpenFile)block_store_file;
-    return 0;
-}
-
-int Longtail_BlockStorage_GetSize(struct Longtail_StorageAPI* storage_api, Longtail_StorageAPI_HOpenFile f, uint64_t* out_size)
-{
-    struct BlockStoreFS* block_store_fs = (struct BlockStoreFS*)storage_api;
-    struct OpenBlockStoreFile* block_store_file = (struct OpenBlockStoreFile*)f;
-    int err = GetAssetSize(block_store_fs, block_store_file, out_size);
-    return err;
-}
-
-int Longtail_BlockStorage_Read(struct Longtail_StorageAPI* storage_api, Longtail_StorageAPI_HOpenFile f, uint64_t offset, uint64_t length, void* output)
-{
-    struct BlockStoreFS* block_store_fs = (struct BlockStoreFS*)storage_api;
-    struct OpenBlockStoreFile* block_store_file = (struct OpenBlockStoreFile*)f;
-    int err = ReadBlockStoreFile(block_store_fs, block_store_file, offset, length, output);
-    return err;
-}
-
-int Longtail_BlockStorage_OpenWriteFile(struct Longtail_StorageAPI* storage_api, const char* path, uint64_t initial_size, Longtail_StorageAPI_HOpenFile* out_open_file)
-{
-    return ENOTSUP;
-}
-
-int Longtail_BlockStorage_Write(struct Longtail_StorageAPI* storage_api, Longtail_StorageAPI_HOpenFile f, uint64_t offset, uint64_t length, const void* input)
-{
-    return ENOTSUP;
-}
-
-int Longtail_BlockStorage_SetSize(struct Longtail_StorageAPI* storage_api, Longtail_StorageAPI_HOpenFile f, uint64_t length)
-{
-    return ENOTSUP;
-}
-
-int Longtail_BlockStorage_SetPermissions(struct Longtail_StorageAPI* storage_api, const char* path, uint16_t permissions)
-{
-    return ENOTSUP;
-}
-
-int Longtail_BlockStorage_GetPermissions(struct Longtail_StorageAPI* storage_api, const char* path, uint16_t* out_permissions)
-{
-    struct BlockStoreFS* block_store_fs = (struct BlockStoreFS*)storage_api;
-    uint64_t path_hash = 0;
-    int err = Longtail_GetPathHash(block_store_fs->hash_api, path, &path_hash);
-    if (err)
-    {
-        return err;
-    }
-    intptr_t asset_index_ptr = hmgeti(block_store_fs->asset_lookup, path_hash);
-    if (asset_index_ptr == -1)
-    {
-        return ENOENT;
-    }
-    uint64_t asset_index = block_store_fs->asset_lookup[asset_index_ptr].value;
-    *out_permissions = block_store_fs->version_index->m_Permissions[asset_index];
-    return 0;
-}
-
-void Longtail_BlockStorage_CloseFile(struct Longtail_StorageAPI* storage_api, Longtail_StorageAPI_HOpenFile f)
-{
-    struct BlockStoreFS* block_store_fs = (struct BlockStoreFS*)storage_api;
-    struct OpenBlockStoreFile* block_store_file = (struct OpenBlockStoreFile*)f;
-    CloseAsset(block_store_fs, block_store_file);
-}
-
-int Longtail_BlockStorage_CreateDir(struct Longtail_StorageAPI* storage_api, const char* path)
-{
-    return ENOTSUP;
-}
-
-int Longtail_BlockStorage_RenameFile(struct Longtail_StorageAPI* storage_api, const char* source_path, const char* target_path)
-{
-    return ENOTSUP;
-}
-
-char* Longtail_BlockStorage_ConcatPath(struct Longtail_StorageAPI* storage_api, const char* root_path, const char* sub_path)
-{
-    size_t root_len = strlen(root_path);
-    if (root_len == 0)
-    {
-        return Longtail_Strdup(sub_path);
-    }
-    if (root_path[root_len - 1] == '/')
-    {
-        --root_len;
-    }
-    size_t sub_len = strlen(sub_path);
-    size_t path_len = root_len + 1 + sub_len;
-    char* path = (char*)Longtail_Alloc(path_len + 1);
-    if (!path)
-    {
-        return 0;
-    }
-    memcpy(path, root_path, root_len);
-    path[root_len] = '/';
-    memcpy(&path[root_len + 1], sub_path, sub_len);
-    return path;
-}
-
-int Longtail_BlockStorage_IsDir(struct Longtail_StorageAPI* storage_api, const char* path)
-{
-    size_t path_len = strlen(path);
-    if (path_len == 0)
-    {
-        return 0;
-    }
-    if (path[strlen(path)- 1] == '/')
-    {
-        return 1;
-    }
-    struct BlockStoreFS* block_store_fs = (struct BlockStoreFS*)storage_api;
-    uint64_t path_hash = 0;
-    int err = Longtail_GetPathHash(block_store_fs->hash_api, path, &path_hash);
-    if (err)
-    {
-        return err;
-    }
-    intptr_t asset_index_ptr = hmgeti(block_store_fs->asset_lookup, path_hash);
-    if (asset_index_ptr != -1)
-    {
-        // Its a file since all paths in version index ends with forward-slash
-        return 0;
-    }
-    char* tmp_path = (char*)Longtail_Alloc(path_len + 1 + 1);
-    strcpy(tmp_path, path);
-    tmp_path[path_len] = '/';
-    tmp_path[path_len + 1] = 0;
-    path_hash = 0;
-    err = Longtail_GetPathHash(block_store_fs->hash_api, path, &path_hash);
-    Longtail_Free(tmp_path);
-    if (err)
-    {
-        return err;
-    }
-    asset_index_ptr = hmgeti(block_store_fs->asset_lookup, path_hash);
-    if (asset_index_ptr == -1)
-    {
-        return 0;
-    }
-    return 1;
-}
-
-int Longtail_BlockStorage_IsFile(struct Longtail_StorageAPI* storage_api, const char* path)
-{
-    size_t path_len = strlen(path);
-    if (path_len == 0)
-    {
-        return 0;
-    }
-    if (path[strlen(path)- 1] == '/')
-    {
-        return 0;
-    }
-    struct BlockStoreFS* block_store_fs = (struct BlockStoreFS*)storage_api;
-    uint64_t path_hash = 0;
-    int err = Longtail_GetPathHash(block_store_fs->hash_api, path, &path_hash);
-    if (err)
-    {
-        return err;
-    }
-    intptr_t asset_index_ptr = hmgeti(block_store_fs->asset_lookup, path_hash);
-    if (asset_index_ptr != -1)
-    {
-        return 1;
-    }
-    return 0;
-}
-
-int Longtail_BlockStorage_RemoveDir(struct Longtail_StorageAPI* storage_api, const char* path)
-{
-    return ENOTSUP;
-}
-
-int Longtail_BlockStorage_RemoveFile(struct Longtail_StorageAPI* storage_api, const char* path)
-{
-    return ENOTSUP;
-}
-
-int Longtail_BlockStorage_StartFind(struct Longtail_StorageAPI* storage_api, const char* path, Longtail_StorageAPI_HIterator* out_iterator)
-{
-    // TODO
-    return ENOTSUP;
-}
-
-int Longtail_BlockStorage_FindNext(struct Longtail_StorageAPI* storage_api, Longtail_StorageAPI_HIterator iterator)
-{
-    // TODO
-    return ENOTSUP;
-}
-
-void Longtail_BlockStorage_CloseFind(struct Longtail_StorageAPI* storage_api, Longtail_StorageAPI_HIterator iterator)
-{
-    // TODO
-}
-
-int Longtail_BlockStorage_GetEntryProperties(struct Longtail_StorageAPI* storage_api, Longtail_StorageAPI_HIterator iterator, struct Longtail_StorageAPI_EntryProperties* out_properties)
-{
-    // TODO
-    return ENOTSUP;
-}
-
-
-void CloseBlockStoreFS(struct Longtail_API* api)
-{
-    struct BlockStoreFS* block_store_fs = (struct BlockStoreFS*)api;
-    hmfree(block_store_fs->asset_lookup);
-    hmfree(block_store_fs->chunk_hash_to_block_index_lookup);
-    Longtail_Free(block_store_fs->chunk_asset_offsets);
-    Longtail_Free(block_store_fs);
-}
-
-int OpenBlockStoreFS(
-    Longtail_HashAPI* hash_api,
-    struct Longtail_JobAPI* job_api,
-    struct Longtail_BlockStoreAPI* block_store,
-    struct Longtail_ContentIndex* content_index,
-    struct Longtail_VersionIndex* version_index,
-    struct BlockStoreFS** out_block_store_fs)
-{
-    struct BlockStoreFS* block_store_fs = (struct BlockStoreFS*)Longtail_Alloc(sizeof(struct BlockStoreFS));
-    block_store_fs->m_API.m_API.Dispose = CloseBlockStoreFS;
-    block_store_fs->hash_api = hash_api;
-    block_store_fs->job_api = job_api;
-    block_store_fs->block_store = block_store;
-    block_store_fs->content_index = content_index;
-    block_store_fs->version_index = version_index;
-    block_store_fs->chunk_hash_to_block_index_lookup = 0;
-    block_store_fs->asset_lookup = 0;
-    block_store_fs->chunk_asset_offsets = (uint64_t*)Longtail_Alloc(sizeof(uint64_t) * (*version_index->m_AssetChunkIndexCount));
-
-    for (uint64_t c = 0; c < *content_index->m_ChunkCount; ++c)
-    {
-        uint64_t block_index = content_index->m_ChunkBlockIndexes[c];
-        hmput(block_store_fs->chunk_hash_to_block_index_lookup, content_index->m_ChunkHashes[c], block_index);
-    }
-    for (uint32_t a = 0; a < *version_index->m_AssetCount; ++a)
-    {
-        TLongtail_Hash path_hash = version_index->m_PathHashes[a];
-        hmput(block_store_fs->asset_lookup, path_hash, a);
-
-        uint32_t chunk_count = block_store_fs->version_index->m_AssetChunkCounts[a];
-        if (chunk_count == 0)
-        {
-            continue;
-        }
-        uint32_t chunk_start_index = block_store_fs->version_index->m_AssetChunkIndexStarts[a];
-        const uint32_t* chunk_indexes = &block_store_fs->version_index->m_AssetChunkIndexes[chunk_start_index];
-        uint64_t* chunk_asset_offsets = &block_store_fs->chunk_asset_offsets[chunk_start_index];
-        uint64_t offset = 0;
-        for (uint32_t c = 0; c < chunk_count; ++c)
-        {
-            uint32_t chunk_index = chunk_indexes[c];
-            uint32_t chunk_size = block_store_fs->version_index->m_ChunkSizes[chunk_index];
-            chunk_asset_offsets[c] = offset;
-            offset += chunk_size;
-        }
-    }
-
-    *out_block_store_fs = block_store_fs;
-    return 0;
-}
-
-void CloseBlockStoreFS(struct BlockStoreFS* block_store_fs)
-{
-    hmfree(block_store_fs->asset_lookup);
-    hmfree(block_store_fs->chunk_hash_to_block_index_lookup);
-    Longtail_Free(block_store_fs->chunk_asset_offsets);
-    Longtail_Free(block_store_fs);
-}
-
-static int BlockStoreStorage_Init(
-    void* mem,
-    Longtail_HashAPI* hash_api,
-    struct Longtail_JobAPI* job_api,
-    struct Longtail_BlockStoreAPI* block_store,
-    struct Longtail_ContentIndex* content_index,
-    struct Longtail_VersionIndex* version_index,
-    struct Longtail_StorageAPI** out_storage_api)
-{
-    return ENOTSUP;
-}
-
-struct Longtail_StorageAPI* Longtail_CreateBlockStoreStorageAPI(
-    Longtail_HashAPI* hash_api,
-    struct Longtail_JobAPI* job_api,
-    struct Longtail_BlockStoreAPI* block_store,
-    struct Longtail_ContentIndex* content_index,
-    struct Longtail_VersionIndex* version_index)
-{
-    LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_INFO, "Longtail_CreateBlockStoreStorageAPI(%p, %p, %p, %p, %p,) failed with %d",
-        hash_api, job_api, block_store, content_index, version_index)
-    LONGTAIL_VALIDATE_INPUT(hash_api != 0, return 0)
-    LONGTAIL_VALIDATE_INPUT(job_api != 0, return 0)
-    LONGTAIL_VALIDATE_INPUT(block_store != 0, return 0)
-    LONGTAIL_VALIDATE_INPUT(content_index != 0, return 0)
-    LONGTAIL_VALIDATE_INPUT(version_index != 0, return 0)
-
-    size_t api_size = sizeof(struct BlockStoreFS);
-    void* mem = Longtail_Alloc(api_size);
-    if (!mem)
-    {
-        LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "Longtail_CreateBlockStoreStorageAPI(%p, %p, %p, %p, %p,) failed with %d",
-            hash_api, job_api, block_store, content_index, version_index,
-            ENOMEM)
-        return 0;
-    }
-    struct Longtail_StorageAPI* storage_api;
-    int err = BlockStoreStorage_Init(
-        mem,
-        hash_api,
-        job_api,
-        block_store,
-        content_index,
-        version_index,
-        &storage_api);
-    if (err)
-    {
-        LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "Longtail_CreateBlockStoreStorageAPI(%p, %p, %p, %p, %p,) failed with %d",
-            hash_api, job_api, block_store, content_index, version_index,
-            err)
-        Longtail_Free(mem);
-        return 0;
-    }
-    return storage_api;
-}
 
 
 TEST(Longtail, TestLongtailBlockFS)
@@ -6958,26 +6186,31 @@ TEST(Longtail, TestLongtailBlockFS)
 
 //    printf("\nReading...\n");
 
-    struct BlockStoreFS* block_store_fs;
-    ASSERT_EQ(0, OpenBlockStoreFS(hash_api, job_api, block_store, block_store_content_index, vindex, &block_store_fs));
+    struct Longtail_StorageAPI* block_store_fs = Longtail_CreateBlockStoreStorageAPI(
+        hash_api,
+        job_api,
+        block_store,
+        block_store_content_index,
+        vindex);
+    ASSERT_NE((struct Longtail_StorageAPI*)0, block_store_fs);
     for (uint32_t f = 0; f < version_paths->m_Count; ++f)
     {
         const char* path = &version_paths->m_PathData[version_paths->m_PathStartOffsets[f]];
         char* full_path = mem_storage->ConcatPath(mem_storage, "source", path);
         if (mem_storage->IsFile(mem_storage, full_path))
         {
-            struct OpenBlockStoreFile* block_store_file;
-            ASSERT_EQ(0, OpenAsset(block_store_fs, path, &block_store_file));
+            Longtail_StorageAPI_HOpenFile block_store_file;
+            ASSERT_EQ(0, block_store_fs->OpenReadFile(block_store_fs, path, &block_store_file));
             uint64_t size;
-            ASSERT_EQ(0, GetAssetSize(block_store_fs, block_store_file, &size));
+            ASSERT_EQ(0, block_store_fs->GetSize(block_store_fs, block_store_file, &size));
             char* buf = (char*)Longtail_Alloc(size);
             if (size > 64)
             {
-                ASSERT_EQ(0, ReadBlockStoreFile(block_store_fs, block_store_file, 0, 32, buf));
-                ASSERT_EQ(0, ReadBlockStoreFile(block_store_fs, block_store_file, 32, 128, &buf[32]));
-                ASSERT_EQ(0, ReadBlockStoreFile(block_store_fs, block_store_file, 160, 96, &buf[160]));
-                ASSERT_EQ(0, ReadBlockStoreFile(block_store_fs, block_store_file, size - 32, 32, &buf[size - 32]));
-                ASSERT_EQ(0, ReadBlockStoreFile(block_store_fs, block_store_file, 256, size - (256 + 32), &buf[256]));
+                ASSERT_EQ(0, block_store_fs->Read(block_store_fs, block_store_file, 0, 32, buf));
+                ASSERT_EQ(0, block_store_fs->Read(block_store_fs, block_store_file, 32, 128, &buf[32]));
+                ASSERT_EQ(0, block_store_fs->Read(block_store_fs, block_store_file, 160, 96, &buf[160]));
+                ASSERT_EQ(0, block_store_fs->Read(block_store_fs, block_store_file, size - 32, 32, &buf[size - 32]));
+                ASSERT_EQ(0, block_store_fs->Read(block_store_fs, block_store_file, 256, size - (256 + 32), &buf[256]));
             }
             else if (size > 0)
             {
@@ -6989,12 +6222,12 @@ TEST(Longtail, TestLongtailBlockFS)
                     {
                         s = size - o;
                     }
-                    ASSERT_EQ(0, ReadBlockStoreFile(block_store_fs, block_store_file, o, s, &buf[o]));
+                    ASSERT_EQ(0, block_store_fs->Read(block_store_fs, block_store_file, o, s, &buf[o]));
                     o += s;
                 }
             }
 
-            CloseAsset(block_store_fs, block_store_file);
+            block_store_fs->CloseFile(block_store_fs, block_store_file);
 
             Longtail_StorageAPI_HOpenFile open_file;
             ASSERT_EQ(0, mem_storage->OpenReadFile(mem_storage, full_path, &open_file));
@@ -7024,7 +6257,7 @@ TEST(Longtail, TestLongtailBlockFS)
         Longtail_Free(full_path);
     }
 
-    CloseBlockStoreFS(block_store_fs);
+    SAFE_DISPOSE_API(block_store_fs);
 
 //    printf("\nDone...\n");
